Upgrade fetch ibus access to IrrevocableIO
[Warning] this change breaks some fetch error handling behavior. The
fix is fetch bookkeeping.

- An assertion adapter is added to check behavior in simulation
- Fetched data is now registered before entering predecode to decouple
  it from fetch ctrl logic. In the future if we pipeline ibus, this can
  be removed.
- The fetcher bookkeeping structure is upgraded from reg to a temporary
  queue to simplify future changes.

This change makes a conv2d kernel ~4% slower so the cycle target is
updated to allow this. This should be resolved once we fully pipeline
the fetch, and we need to reset cycle targets for all similar tests.

Bypass-Presubmit-Reason: To be submitted together with the next change
which fixes fetch error behavior

Change-Id: I6dfa3ea1e820495111908f54f69c3f9ad5dd4a26